initial field decoding support
Because field naming is not uniform (a CNAM in the TES4 record is a
different format than one in other records) decoding them isn't as
straightforward as it initially looks. The decoders are named in the format
"{record type}.{field type}" (TES4.CNAM for example). Because some records
are uniform (I'm looking at you EDID), we also support generic decoders. If
no specific decoder is found we will instead look for a generic one. If
neither is found we print out an MD5 of the data.